Add swift file watcher to catch when a test file has been edited
and test discovery needs to be run again
